% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_functions.R
\name{corner}
\alias{corner}
\title{Corner}
\usage{
corner(df)
}
\arguments{
\item{df}{Dataframe, matrix or tibble to be shown}
}
\description{
Getting the top left corner of a dataframe, tibble or matrix.
}
