% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_functions.R
\name{filter.expressed.genes}
\alias{filter.expressed.genes}
\title{Filters out lowly expressed genes}
\usage{
\method{filter}{expressed.genes}(sobj, min_cells_expressed = 25)
}
\arguments{
\item{sobj}{Seurat object}

\item{min_cells_expressed}{In how many cells needs a gene being expressed to be accepted}
}
\description{
Remove genes from the dataset which are lowly expressed.
}
