% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_functions.R
\name{filter.nuc.droplets}
\alias{filter.nuc.droplets}
\title{Filtering droplets}
\usage{
filter.nuc.droplets(sobj.exon, min.frac.intronic = 0.5,
  max.rank = NULL, max.frac.intronic = NULL, min.rank = NULL)
}
\arguments{
\item{sobj.exon}{Seurat object, pretreated with generate.frac.intron.meta}
}
\description{
Filters droplets based on their properties, calculated using generate.frac.intron.meta
}
