% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulated_functions.R
\name{hdbscan_clustering}
\alias{hdbscan_clustering}
\title{Find clusters using HDBSCAN}
\usage{
hdbscan_clustering(adata)
}
\arguments{
\item{adata}{Seurat object with var.genes slot filled.}
}
\value{
A character vector of class assignments. The class "-1" is used to label cells which could not be assigned to a cluster.
}
\description{
This function will use the scale.data slot of the Seurat object to find clusters using the HDBSCAN algorithm.
HDBSCAN is implemented in python and connected to R using the reticulate package.
}
\details{
Make sure you have installed reticulate and enabled the right environment, before running this function.
The following python (3.6) packages need to be installed:

  - umap

  - hdbscan
}
