% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_functions.R
\name{make.seurat.object}
\alias{make.seurat.object}
\title{Generates a seurat object of 10x raw data}
\usage{
make.seurat.object(path, cell.name.addition = NULL, min.genes = 100)
}
\arguments{
\item{path}{Path to the 10x output}

\item{cell.name.addition}{Prefix for each cell barcode}

\item{min.genes}{Minimal number of genes which need to be detected per barcode to keep the barcode}
}
\value{
Seurat object
}
\description{
Using the 10x output path with the genes, barcodes and matrix files.
}
