% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_functions.R
\name{export.csvs}
\alias{export.csvs}
\title{Exports a seurat object as CSV files}
\usage{
export.csvs(obj, folderpath, sample.name = "seurat",
  alter.gene.names = NULL)
}
\arguments{
\item{obj}{Seurat object}

\item{folderpath}{Folder to export to. If not existent, it will be created.}

\item{sample.name}{Prefix for each exported CSV}

\item{alter.gene.names}{Dataframe to convert the names of the Genes.}
}
\description{
This function creates a CSV file for:
  - the data field
  - the raw.data field
  - the UMAP embedding
}
